Creating an event and method
I have created a button click event for button 7 and also a listbox
which contains the sequences for each charater button 7 will go through
when clicked by the user. After I created these, I have procceeded to
create a method which allows the button click event, to pass which
button is being clicked, which listbox that is linked to and also a
specififed number for that button event.