Button 8
I have now created a new listBox with the variable name listBoxKey8 and
also changed the variable name of the button 8 to btn8. After this I
have created at button click event for button 8, and then implemented
the same functionality into button 7. I finally tested it to see if it
worked and it did.