Button 4
I have created another listBox with the variable name listBoxKey4 and
also changed the variable name of the button to btn4. After this I again
created a button click event and implemented the same functionality from
button 7, and linking to my method called called button response. I
finally, tested to see if it worked and there was no errors.