Button 5
I have created another listBox with the variable name listBoxKey5 and
also changed the variable name of the button to btn5. After this I
created a button click event and implemented the same functionality I
added in the previous buttons, giving it a specific button case number.
(this is a variable i set for each button as it specifically identifys
that button. I added this into the other buttons aswell.).