Button 6
I again created another listBox, this time with the variable name
listBoxKey6 and also changed the variable name of the button to btn6.
After this, I created a button click event, and implemented the same
functionality from the other button click events. I finally test to see
if the button worked with the method button response and the correct
sequence was displayed.