Button 1
I have now implemnted in another listBox and given it the variable name
listBoxKey1 and also changed the name of the button to btn1. After this
I created a button click event on btn1 and implemented the same
functionality, that I did with the other button click events. I finally,
tested to see if it worked and the result was it displayed the right
sequence from the listbox.