Button 2
I created another listBox and changed the variable name to listBoxKey2
and also changed the vaiable name of button2 to btn2. Afterwards. I
again created another button click event. and implemented the same
functionality as I did with the other buttons previously. Finally, I
tests the button to see if it would display the correct sequence from
the listBox and it correctly did.