Button 3
I have again created another listBox this time with the variable name
listBoxKey3 and also changed the variable name of button3 to btn3. After
this, I created a button click event for this button, and added the
correct functionality for it too link with the button response method,
to give the correct sequence outcome. Finally, I tested this, and the
result was that the button displayed the correct sequence to output.