Button Asterisk
I have created a listBox with the variable name listBoxKeyAsterisk and
also changed the name of button 10 to btnAsterisk. After this, I created
a button click event and added in the same functionality as the other
buttons, passing variables to the button repsonse method, to gain the
correct sequence to be displayed. Finally, I tested this button to see
if it gave the correct sequence output and it did so.