Exception Error
While creating the functionality for button 0, I ran into an error being
that when clicking on the button too quickly after a character button it
would make the program crash. Therefore, I used the debugger to find
that the exception was occuring in the timer method, therefore I used a
try and catch method to catch the exception when it happens and ignore
it.