Button Enter
I changed the variable name of the enter key to btnEnter and created a
click event. I then proceeded to use the variable txtOutput and use
Environment.NewLine to create a new line every time it was clicked. I
then procceed to check if it works and it does without fault.