Configure menu button
I created a click event for the configure button and implement
functionality for it to work correctly. Firstly, I created an int being
int userInput to get the users input which would be used to change the
value of the timers interval. I link my class to the myDialogs class, so
I could access the InputBox method.