Save menu option
I have added into my class the using System.IO function which allows me
to use StreamWriter and File.CreateText. First I changed the variable
name of the menu button to menuSave and then created a button click
event. Afterwards to procceeded to add functionality, such as setting a
variable called streamWriter which would be used to write, open and save
as.