Open menu option
I firstly created a fileOpenDialog using the toolbox and changed its
variable name to openFileD. I then changed the variable name of the menu
open button to menuOpen. After this I created a button click event,
which I would then add functionality in for it to work effectively. I
used the variable Bool_Saving to see if the program needs saving first.
Then I would use the vaiable openFileD to give the user a dialog in
which they can find the file they are trying to access. I then used the
variable strReader to be used to store the text already saved on the
file and then outputed into the word viewer on the program.