000 button modification
I realised I had not implemented into my code, that when the 000 button
has been clicked and it appends the text in the word builder textbox
that it clears the variable Str_KeyStrokes. Therefore, I proceeded to
create a text box which I gave the variable name txtBoxKeyStrokes. I
then at the end of the method for 000 button click event I set
Str_KeyStrokes to = txtBoxKeyStrokes.Text which cleared the variable.